// apps/app/src/components/design-system/research-ui/documents/reference-prototype/types/documents.ts
export interface VintagePosterDProps {
  title?: string
  avatar?: string
  avatarLabel?: string
  heroImage?: string
  colA?: string
  colB?: string
  colC?: string
  stampText?: string
}